+++ app/preload/index.js
@@ -24,10 +24,11 @@ require("./follow")
 
 const {pathToSpecialPageName} = require("../util")
 const specialPage = pathToSpecialPageName(window.location.href)
+const skipProtocols = ["sourceviewer:", "readerview:", "markdownviewer:"]
 if (specialPage.name) {
     // Load the special page specific JavaScript
     require(`./${specialPage.name}`)
-} else {
+} else if (!skipProtocols.find(p => window.location.href.startsWith(p))) {
     // Load the failed page information handler for nonspecial pages
     require("./failedload")
     // Load the local directory browser for nonspecial pages