+++ app/renderer/actions.js
@@ -242,6 +242,43 @@ const toggleReaderViewNewTab = () => {
     }
 }
 
+const toggleMarkdownViewer = () => {
+    const {navigateTo} = require("./tabs")
+    if (currentPage().src.startsWith("markdownviewer:")) {
+        const src = currentPage().src.replace(/^markdownviewer:\/?\/?/g, "")
+        let loc = String(src)
+        if (process.platform !== "win32" && !loc.startsWith("/")) {
+            loc = `/${loc}`
+        }
+        if (isFile(loc) || isDir(loc)) {
+            navigateTo(`file://${loc}`)
+            return
+        }
+        navigateTo(`https://${src}`)
+    } else {
+        navigateTo(currentPage().src.replace(/^.+?:\/?\/?/g, "markdownviewer:"))
+    }
+}
+
+const toggleMarkdownViewerNewTab = () => {
+    const {navigateTo, addTab} = require("./tabs")
+    if (currentPage().src.startsWith("markdownviewer:")) {
+        const src = currentPage().src.replace(/^markdownviewer:\/?\/?/g, "")
+        let loc = String(src)
+        if (process.platform !== "win32" && !loc.startsWith("/")) {
+            loc = `/${loc}`
+        }
+        if (isFile(loc) || isDir(loc)) {
+            navigateTo(`file://${loc}`)
+            return
+        }
+        navigateTo(`https://${src}`)
+    } else {
+        addTab({"url": currentPage().src.replace(
+            /^.+?:\/?\/?/g, "markdownviewer:")})
+    }
+}
+
 const toExploreMode = () => {
     const {setMode} = require("./modes")
     setMode("explore")
@@ -1037,6 +1074,8 @@ module.exports = {
     toTopSplitWindow,
     toggleAlwaysOnTop,
     toggleFullscreen,
+    toggleMarkdownViewer,
+    toggleMarkdownViewerNewTab,
     toggleReaderView,
     toggleReaderViewNewTab,
     toggleSourceViewer,